refactor: refine app structures
