+++ src/roots.ts
@@ -1,0 +1,8 @@
+import { Hono } from "hono";
+
+export const rootRoute = new Hono().get("/", (c) => {
+  return c.json({
+    message: "School API",
+    schools: "/schools",
+  });
+});
